+++ zetaclient/chains/sui/observer/observer.go
@@ -74,7 +74,7 @@ func (ob *Observer) CheckRPCStatus(ctx context.Context) error {
 	return nil
 }
 
-// PostGasPrice posts Sui gas price to zetacore.
+// ObserveGasPrice posts Sui gas price to zetacore.
 // Note (1) that Sui changes gas per EPOCH (not block)
 // Note (2) that SuiXGetCurrentEpoch() is deprecated.
 //
@@ -86,13 +86,13 @@ func (ob *Observer) CheckRPCStatus(ctx context.Context) error {
 // - "During regular network usage, users are NOT expected to pay tips"
 // - "Validators update the ReferencePrice every epoch (~24h)"
 // - "Storage price is updated infrequently through gov proposals"
-func (ob *Observer) PostGasPrice(ctx context.Context) error {
+func (ob *Observer) ObserveGasPrice(ctx context.Context) error {
 	checkpoint, err := ob.suiClient.GetLatestCheckpoint(ctx)
 	if err != nil {
 		return errors.Wrap(err, "unable to get latest checkpoint")
 	}
 
-	epochNum, err := uint64FromStr(checkpoint.Epoch)
+	epoch, err := uint64FromStr(checkpoint.Epoch)
 	if err != nil {
 		return errors.Wrap(err, "unable to parse epoch number")
 	}
@@ -104,12 +104,13 @@ func (ob *Observer) PostGasPrice(ctx context.Context) error {
 		return errors.Wrap(err, "unable to get ref gas price")
 	}
 
-	// no priority fee for Sui
+	// There's no concept of priority fee in Sui.
 	const priorityFee = 0
 
-	_, err = ob.ZetacoreClient().PostVoteGasPrice(ctx, ob.Chain(), gasPrice, priorityFee, epochNum)
+	logger := ob.Logger().Chain
+	_, err = ob.ZetaRepo().VoteGasPrice(ctx, logger, gasPrice, priorityFee, epoch)
 	if err != nil {
-		return errors.Wrap(err, "unable to post vote for gas price")
+		return err
 	}
 
 	ob.setLatestGasPrice(gasPrice)