+++ zetaclient/chains/zrepo/client.go
@@ -14,11 +14,15 @@ import (
 	fungible "github.com/zeta-chain/node/x/fungible/types"
 	observer "github.com/zeta-chain/node/x/observer/types"
 	keys "github.com/zeta-chain/node/zetaclient/keys/interfaces"
+	"github.com/zeta-chain/node/zetaclient/zetacore"
 )
+
+var _ ZetacoreClient = &zetacore.Client{}
 
 type ChainID = int64
 type Nonce = uint64
 
+// ZetacoreWriter contains the functions that mutate ZetaChain state.
 type ZetacoreWriter interface {
 	PostVoteGasPrice(_ context.Context,
 		_ chains.Chain,
@@ -58,39 +62,17 @@ type ZetacoreWriter interface {
 	) (string, error)
 }
 
-// ZetacoreClient is the client interface that interacts with zetacore.
-//
-//go:generate mockery --name ZetacoreClient --filename zetacore_client.go --case underscore --output ../../testutils/mocks
-type ZetacoreClient interface {
+// ZetacoreClientRepo contains the functions used by ZetaRepo.
+type ZetacoreClientRepo interface {
 	ZetacoreWriter
 
 	Chain() chains.Chain
 
 	GetKeys() keys.ObserverKeys
 
-	GetSupportedChains(context.Context) ([]chains.Chain, error)
-
-	GetAdditionalChains(context.Context) ([]chains.Chain, error)
-
-	GetChainParams(context.Context) ([]*observer.ChainParams, error)
-
-	GetForeignCoinsFromAsset(context.Context, ChainID, eth.Address) (fungible.ForeignCoins, error)
-
-	GetKeyGen(context.Context) (observer.Keygen, error)
-
-	GetTSS(context.Context) (observer.TSS, error)
-	GetTSSHistory(context.Context) ([]observer.TSS, error)
-
-	GetBlockHeight(context.Context) (int64, error)
-
 	ListPendingCCTX(context.Context, chains.Chain) ([]*crosschain.CrossChainTx, uint64, error)
 
-	ListPendingCCTXWithinRateLimit(context.Context,
-	) (*crosschain.QueryListPendingCctxWithinRateLimitResponse, error)
-
-	GetRateLimiterInput(_ context.Context,
-		window int64,
-	) (*crosschain.QueryRateLimiterInputResponse, error)
+	GetForeignCoinsFromAsset(context.Context, ChainID, eth.Address) (fungible.ForeignCoins, error)
 
 	GetPendingNoncesByChain(context.Context, ChainID) (observer.PendingNonces, error)
 
@@ -106,19 +88,49 @@ type ZetacoreClient interface {
 
 	GetInboundTrackersForChain(context.Context, ChainID) ([]crosschain.InboundTracker, error)
 
+	NewBlockSubscriber(context.Context) (chan cometbft.EventDataNewBlock, error)
+
+	GetBTCTSSAddress(context.Context, ChainID) (string, error)
+}
+
+// ZetacoreClient is the client interface that interacts with zetacore.
+//
+// TODO: this should be moved elsewhere, since it is not used by ZetaRepo.
+// See: https://github.com/zeta-chain/node/issues/4300
+//
+//go:generate mockery --name ZetacoreClient --filename zetacore_client.go --case underscore --output ../../testutils/mocks
+type ZetacoreClient interface {
+	ZetacoreClientRepo
+
+	GetSupportedChains(context.Context) ([]chains.Chain, error)
+
+	GetAdditionalChains(context.Context) ([]chains.Chain, error)
+
+	GetChainParams(context.Context) ([]*observer.ChainParams, error)
+
+	GetKeyGen(context.Context) (observer.Keygen, error)
+
+	GetTSS(context.Context) (observer.TSS, error)
+	GetTSSHistory(context.Context) ([]observer.TSS, error)
+
+	GetBlockHeight(context.Context) (int64, error)
+
+	ListPendingCCTXWithinRateLimit(context.Context,
+	) (*crosschain.QueryListPendingCctxWithinRateLimitResponse, error)
+
+	GetRateLimiterInput(_ context.Context,
+		window int64,
+	) (*crosschain.QueryRateLimiterInputResponse, error)
+
 	GetCrosschainFlags(context.Context) (observer.CrosschainFlags, error)
 	GetRateLimiterFlags(context.Context) (crosschain.RateLimiterFlags, error)
 	GetOperationalFlags(context.Context) (observer.OperationalFlags, error)
 
 	GetObserverList(context.Context) ([]string, error)
 
-	GetBTCTSSAddress(context.Context, ChainID) (string, error)
-
 	GetZetaHotKeyBalance(context.Context) (cosmosmath.Int, error)
 
 	GetUpgradePlan(context.Context) (*upgrade.Plan, error)
-
-	NewBlockSubscriber(context.Context) (chan cometbft.EventDataNewBlock, error)
 
 	GetBaseGasPrice(context.Context) (int64, error)